Revert last commit + Add NTR mode switch
This no longer needs dev launcher patch to work. Reverted last commit
because MPU change broke akloader when attempting to boot a retail rom.

Touch screen is not working in the menu (this does not impact games
however). Related to new mode switch. A force DS touchscreen function is
needed for this to be fixed.

The touch screen works fine however if you boot it from ntr mode hbmenu
or NTR Bootstrap.